feat: replaced podcast player with socials
